Convert Risk Bubble to an ES6 module
