+++ ui/legacy.js
@@ -9,7 +9,6 @@ import './legacyHelpers';
 // student profile page:
   // pure ui components:
 import '../app/assets/javascripts/student_profile/quad_converter.jsx';
-import '../app/assets/javascripts/student_profile/risk_bubble.jsx';
 import '../app/assets/javascripts/student_profile/service_color.jsx';
 import '../app/assets/javascripts/student_profile/provided_by_educator_dropdown.jsx';
 import '../app/assets/javascripts/student_profile/profile_chart_settings.jsx';
